Remove single jurisdiction backend (#830)
* Remove single-jurisdiction flow from backend

Also changes the /election/new endpoint to just /election.

* Remove Election.is_multi_jurisdiction and other unused fields

* Remove server test parallelism

Race condition errors seem to be more common after these changes.
