+++ client/src/components/HomeScreen.test.tsx
@@ -16,7 +16,7 @@ const apiCalls = {
     },
   },
   postNewAudit: (body: {}) => ({
-    url: '/api/election/new',
+    url: '/api/election',
     options: {
       method: 'POST',
       body: JSON.stringify(body),
@@ -40,7 +40,6 @@ const apiCalls = {
               auditName: 'November Presidential Election 2020',
               electionName: '',
               state: 'CA',
-              isMultiJurisdiction: true,
             },
           ],
         },
@@ -61,7 +60,6 @@ const apiCalls = {
               auditName: 'November Presidential Election 2020',
               electionName: '',
               state: 'CA',
-              isMultiJurisdiction: true,
             },
           ],
         },
@@ -113,7 +111,6 @@ describe('Home screen', () => {
         organizationId: 'org-id',
         auditName: 'November Presidential Election 2020',
         auditType: 'BATCH_COMPARISON',
-        isMultiJurisdiction: true,
       }),
       ...setupScreenCalls,
       aaApiCalls.getJurisdictionFile,
@@ -185,7 +182,6 @@ describe('Home screen', () => {
         organizationId: 'org-id-2',
         auditName: 'Presidential Primary',
         auditType: 'BALLOT_POLLING',
-        isMultiJurisdiction: true,
       }),
       ...setupScreenCalls,
       aaApiCalls.getJurisdictionFile,
