+++ server/auth/routes.py
@@ -73,7 +73,6 @@ def serialize_election(election):
         "auditName": election.audit_name,
         "electionName": election.election_name,
         "state": election.state,
-        "isMultiJurisdiction": election.is_multi_jurisdiction,
     }
