+++ server/bgcompute.py
@@ -1,10 +1,8 @@
 import time
-from sqlalchemy import update
 
 from server.app import app
 from server.database import db_session
 from server.models import *  # pylint: disable=wildcard-import
-from server.api.routes import compute_sample_sizes
 from server.util.jurisdiction_bulk_update import process_jurisdictions_file
 from server.api.standardized_contests import process_standardized_contests_file
 from server.api.ballot_manifest import process_ballot_manifest_file
@@ -13,61 +11,11 @@ from server.api.cvrs import process_cvr_file
 
 
 def bgcompute():
-    bgcompute_compute_round_contests_sample_sizes()
     bgcompute_update_election_jurisdictions_file()
     bgcompute_update_standardized_contests_file()
     bgcompute_update_ballot_manifest_file()
     bgcompute_update_batch_tallies_file()
     bgcompute_update_cvr_file()
-
-
-def bgcompute_compute_round_contests_sample_sizes():
-    # Round contests that don't have sample_size_options - only in single-jurisdiction audits
-    # Multi-jurisdiction audits compute estimated sample sizes on demand
-    round_contests = (
-        RoundContest.query.filter_by(sample_size_options=None)
-        .join(Round)
-        .join(Election)
-        .filter_by(is_multi_jurisdiction=False)
-        .all()
-    )
-
-    for round_contest in round_contests:
-        try:
-            app.logger.info(
-                "computing sample size options for round {:d} of election ID {:s}".format(
-                    round_contest.round.round_num, round_contest.round.election_id
-                )
-            )
-
-            # Claim this RoundContest by updating the sample_size_options field
-            # to an "in progress" value This acts like a lock on the
-            # RoundContest so it can only get processed once (since this
-            # process has a side effect of sampling ballots in some cases,
-            # which we only want to happen once).
-            result = db_session.execute(
-                update(RoundContest.__table__)  # pylint: disable=no-member
-                .where(RoundContest.contest_id == round_contest.contest_id)
-                .where(RoundContest.round_id == round_contest.round_id)
-                .where(RoundContest.sample_size_options.is_(None))
-                .values(sample_size_options="CALCULATION_IN_PROGRESS")
-            )
-            if result.rowcount == 0:
-                continue
-
-            compute_sample_sizes(round_contest)
-
-            app.logger.info(
-                "done computing sample size options for round {:d} of election ID {:s}: {:s}".format(
-                    round_contest.round.round_num,
-                    round_contest.round.election_id,
-                    round_contest.sample_size_options,
-                )
-            )
-        except Exception:
-            app.logger.exception(
-                "ERROR while computing sample size options, continuing to next one."
-            )
 
 
 def bgcompute_update_election_jurisdictions_file() -> int:
