--- server/tests/single_jurisdiction/manifest.csv
@@ -1,485 +1,0 @@
-Batch Name,Number of Ballots,Container,Tabulator
-1,200,Red,4
-2,200,Red,4
-3,200,Red,3
-4,200,Red,3
-5,201,Red,4
-6,200,Red,6
-7,113,Red,6
-8,200,Red,6
-9,200,Red,1
-10,200,Red,1
-11,200,Red,1
-12,200,Red,1
-13,200,Red,4
-14,200,Red,4
-15,200,Red,4
-16,200,Red,4
-17,200,Red,3
-18,200,Red,3
-19,200,Red,3
-20,200,Red,3
-21,200,Red,6
-22,164,Red,6
-23,200,Red,6
-24,200,Red,6
-25,200,Red,4
-26,200,Red,4
-27,200,Red,4
-28,200,Red,4
-29,200,Red,3
-30,200,Red,3
-31,200,Red,3
-32,50,Red,3
-33,200,Red,1
-34,200,Red,1
-35,200,Red,1
-36,199,Red,1
-37,199,Red,6
-38,200,Red,6
-39,200,Red,6
-40,137,Red,6
-41,200,Red,4
-42,200,Red,4
-43,200,Red,4
-44,200,Red,4
-45,200,Red,3
-46,146,Red,3
-47,200,Red,3
-48,200,Red,3
-49,200,Red,6
-50,200,Red,6
-51,200,Red,6
-52,200,Red,6
-53,162,Red,1
-54,200,Red,1
-55,200,Red,1
-56,200,Red,1
-57,200,Red,6
-58,200,Red,6
-59,200,Red,6
-60,200,Red,6
-61,200,Red,4
-62,200,Red,4
-63,200,Red,4
-64,200,Red,4
-65,200,Red,3
-66,158,Red,3
-67,200,Red,3
-68,199,Red,3
-69,200,Red,4
-70,200,Red,4
-71,200,Red,4
-72,200,Red,4
-73,200,Red,1
-74,38,Red,1
-75,199,Red,1
-76,200,Red,1
-77,200,Red,3
-78,200,Red,3
-79,199,Red,3
-80,200,Red,3
-81,200,Red,6
-82,200,Red,6
-83,200,Red,6
-84,152,Red,6
-85,200,Red,1
-86,200,Red,1
-87,200,Red,1
-88,200,Red,1
-89,199,Red,4
-90,200,Red,4
-91,200,Red,4
-92,201,Red,4
-93,200,Red,3
-94,200,Red,3
-95,200,Red,3
-96,200,Red,3
-97,200,Red,6
-98,135,Red,6
-99,200,Red,6
-100,200,Red,6
-101,199,Red,3
-102,200,Red,3
-103,200,Red,3
-104,200,Red,3
-105,98,Red,4
-106,200,Red,4
-107,199,Red,4
-108,200,Red,4
-109,199,Red,1
-110,200,Red,1
-111,200,Red,1
-112,200,Red,1
-113,200,Red,6
-114,200,Red,6
-115,200,Red,6
-116,200,Red,6
-117,199,Red,4
-118,113,Red,4
-119,200,Red,4
-120,200,Red,4
-121,200,Red,3
-122,200,Red,3
-123,200,Red,3
-124,106,Red,3
-125,200,Red,1
-126,200,Red,1
-127,200,Red,1
-128,200,Red,1
-129,201,Red,6
-130,199,Red,6
-131,200,Red,6
-132,200,Red,6
-133,200,Red,4
-134,46,Red,4
-135,200,Red,4
-136,200,Red,4
-137,199,Red,3
-138,199,Red,3
-139,200,Red,3
-140,200,Red,3
-141,200,Red,6
-142,200,Red,6
-143,200,Red,6
-144,200,Red,6
-145,200,Red,4
-146,200,Red,4
-147,173,Red,4
-148,200,Red,4
-149,201,Red,1
-150,200,Red,1
-151,200,Red,1
-152,200,Red,1
-153,8,Red,3
-154,200,Red,3
-155,200,Red,3
-156,200,Red,3
-157,200,Red,6
-158,200,Red,6
-159,112,Red,6
-160,199,Red,6
-161,200,Red,6
-162,200,Red,6
-163,200,Red,6
-164,200,Red,6
-165,200,Red,4
-166,199,Red,4
-167,200,Red,4
-168,200,Red,4
-169,176,Red,3
-170,200,Red,3
-171,41,Red,3
-172,200,Red,3
-173,200,Red,4
-174,200,Red,4
-175,200,Red,4
-176,128,Red,4
-177,201,Red,4
-178,199,Red,3
-179,95,Red,5
-180,200,Red,5
-181,200,Red,1
-182,199,Red,1
-183,200,Red,1
-184,200,Red,1
-185,200,Red,6
-186,199,Red,6
-187,200,Red,6
-188,200,Red,6
-189,200,Red,4
-190,199,Red,4
-191,192,Red,4
-192,200,Red,4
-193,200,Red,6
-194,200,Red,6
-195,198,Red,6
-196,200,Red,6
-197,200,Red,1
-198,144,Red,1
-199,201,Red,1
-200,200,Red,1
-201,200,Red,4
-202,200,Red,4
-203,200,Red,4
-204,200,Red,4
-205,200,Red,3
-206,72,Red,5
-207,200,Red,5
-208,200,Red,5
-209,200,Red,6
-210,199,Red,6
-211,199,Red,6
-212,201,Red,6
-213,200,Red,1
-214,200,Red,1
-215,200,Red,1
-216,152,Red,1
-217,200,Red,4
-218,5,Red,4
-219,195,Red,4
-220,199,Red,4
-221,200,Red,5
-222,197,Red,5
-223,200,Red,5
-224,200,Red,4
-225,199,Red,1
-226,103,Red,1
-227,200,Red,1
-228,200,Red,1
-229,200,Red,6
-230,200,Red,6
-231,200,Red,6
-232,199,Red,6
-233,99,Red,6
-234,45,Red,6
-235,200,Red,1
-236,199,Red,1
-237,200,Red,1
-238,200,Red,3
-239,200,Red,1
-240,199,Red,1
-241,200,Red,1
-242,200,Red,3
-243,200,Red,3
-244,200,Red,4
-245,200,Red,4
-246,200,Red,4
-247,200,Red,4
-248,199,Red,4
-249,200,Red,1
-250,199,Red,1
-251,199,Red,1
-252,200,Red,1
-253,144,Red,4
-254,199,Red,4
-255,184,Red,4
-256,200,Red,4
-257,200,Red,1
-258,200,Red,1
-259,200,Red,1
-260,199,Red,1
-261,200,Red,5
-262,200,Red,5
-263,200,Red,5
-264,200,Red,5
-265,142,Red,4
-266,200,Red,4
-267,200,Red,4
-268,200,Red,4
-269,200,Red,1
-270,200,Red,1
-271,200,Red,1
-272,200,Red,1
-273,199,Red,4
-274,67,Red,4
-275,200,Red,4
-276,200,Red,4
-277,200,Red,1
-278,200,Red,1
-279,200,Red,1
-280,199,Red,1
-281,199,Red,4
-282,200,Red,4
-283,200,Red,4
-284,127,Red,4
-285,200,Red,1
-286,200,Red,1
-287,200,Red,1
-288,200,Red,1
-289,200,Red,4
-290,199,Red,4
-291,200,Red,4
-292,200,Red,4
-293,200,Red,1
-294,154,Red,1
-295,200,Red,1
-296,182,Red,1
-297,200,Red,1
-298,199,Red,1
-299,200,Red,1
-300,200,Red,1
-301,200,Red,1
-302,200,Red,1
-303,200,Red,1
-304,200,Red,1
-305,199,Red,1
-306,139,Red,1
-307,200,Red,4
-308,200,Red,4
-309,54,Red,1
-310,186,Red,4
-311,199,Red,1
-312,200,Red,1
-313,199,Red,4
-314,201,Red,4
-315,199,Red,4
-316,198,Red,4
-317,199,Red,1
-318,200,Red,1
-319,200,Red,1
-320,149,Red,1
-321,198,Red,4
-322,200,Red,4
-323,200,Red,4
-324,200,Red,4
-325,200,Red,1
-326,200,Red,1
-327,192,Red,1
-328,200,Red,1
-329,200,Red,5
-330,199,Red,5
-331,200,Red,5
-332,200,Red,5
-333,200,Red,6
-334,138,Red,6
-335,200,Red,6
-336,200,Red,6
-337,200,Red,2
-338,200,Red,2
-339,199,Red,2
-340,126,Red,2
-341,200,Red,4
-342,200,Red,4
-343,200,Red,4
-344,200,Red,4
-345,200,Red,1
-346,199,Red,1
-347,200,Red,1
-348,199,Red,1
-349,200,Red,6
-350,120,Red,6
-351,200,Red,6
-352,200,Red,6
-353,200,Red,2
-354,199,Red,2
-355,200,Red,2
-356,201,Red,2
-357,200,Red,5
-358,200,Red,5
-359,200,Red,5
-360,138,Red,5
-361,200,Red,1
-362,199,Red,1
-363,200,Red,1
-364,200,Red,1
-365,200,Red,4
-366,200,Red,4
-367,199,Red,4
-368,139,Red,4
-369,200,Red,5
-370,200,Red,5
-371,200,Red,5
-372,200,Red,5
-373,54,Red,6
-374,200,Red,6
-375,200,Red,6
-376,200,Red,6
-377,200,Red,6
-378,200,Red,6
-379,200,Red,6
-380,200,Red,6
-381,200,Red,2
-382,200,Red,2
-383,200,Red,2
-384,200,Red,2
-385,154,Red,1
-386,200,Red,1
-387,200,Red,1
-388,200,Red,1
-389,200,Red,4
-390,200,Red,4
-391,200,Red,4
-392,200,Red,4
-393,200,Red,5
-394,200,Red,5
-395,143,Red,5
-396,199,Red,5
-397,200,Red,5
-398,200,Red,5
-399,43,Red,5
-400,200,Red,1
-401,200,Red,2
-402,200,Red,2
-403,200,Red,2
-404,200,Red,2
-405,200,Red,4
-406,200,Red,4
-407,200,Red,4
-408,200,Red,4
-409,142,Red,4
-410,200,Red,5
-411,200,Red,5
-412,200,Red,5
-413,200,Red,5
-414,200,Red,4
-415,200,Red,4
-416,199,Red,1
-417,199,Red,5
-418,199,Red,5
-419,99,Red,5
-420,200,Red,5
-421,200,Red,1
-422,200,Red,1
-423,200,Red,1
-424,200,Red,1
-425,200,Red,6
-426,200,Red,1
-427,200,Red,1
-428,200,Red,1
-429,137,Red,6
-430,200,Red,4
-431,200,Red,4
-432,200,Red,4
-433,199,Red,5
-434,200,Red,5
-435,200,Red,5
-436,199,Red,5
-437,200,Red,6
-438,200,Red,6
-439,154,Red,6
-440,200,Red,6
-441,200,Red,2
-442,200,Red,2
-443,199,Red,2
-444,200,Red,2
-445,200,Red,1
-446,118,Red,4
-447,199,Red,4
-448,63,Red,4
-449,9,Red,4
-450,12,Red,4
-451,6,Red,4
-452,8,Red,4
-453,11,Red,4
-454,3,Red,4
-455,6,Red,4
-456,6,Red,4
-457,11,Red,4
-458,9,Red,4
-459,5,Red,4
-460,8,Red,4
-461,12,Red,4
-462,7,Red,4
-463,11,Red,4
-464,23,Red,4
-465,11,Red,4
-466,15,Red,4
-467,69,Red,4
-468,19,Red,4
-469,11,Red,4
-470,3,Red,4
-471,22,Red,4
-472,120,Red,4
-473,66,Red,4
-474,49,Red,4
-475,36,Red,4
-476,27,Red,4
-477,45,Red,4
-478,24,Red,4
-479,21,Red,4
-480,32,Red,4
-481,221,Red,3
-482,82,Red,3
-483,71,Red,4
-484,66,Red,4
