--- server/tests/single_jurisdiction/small-manifest.csv
@@ -1,11 +1,0 @@
-Batch Name,Number of Ballots
-1,117
-2,269
-3,178
-4,213
-5,234
-6,253
-7,120
-8,244
-9,355
-10,134
